requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
python-dotenv==1.0.1
telethon==1.32.0
beautifulsoup4==4.12.3
requests==2.31.0